Add GPT library for generating commit messages
This commit adds the GPT library to generate commit messages for Git commits. The `git-commit` command now uses this library to generate commit messages instead of using the `args` parameter. The library is also used in the `github_c` package for generating commit messages when pushing to GitHub repositories. Additionally, this commit includes minor changes to fix formatting and improve readability of log messages.