Refactor git commit command and update gitignore
The git commit command in the git-commit.go file has been refactored to include the update of the Go module and the creation of a gitignore file. The git.GitignoreCreate(directory) line has been commented out temporarily. The commit also includes the update of the gitignore.txt file, removing unnecessary patterns and adding patterns for Svelte and Angular projects.