Fix git repository detection and update clog package version
The commit fixes an issue where the git repository was not being detected correctly. The error message has also been updated to use the `clog.Errorf` function for consistency.

The commit also updates the version of the `clog` package to `v0.0.0-20230930052527-683e18a3d08d` in the `go.mod` file.